requirements/base.txt
# Base requirements for installation *except* Django
#
# Isolating Django from the other requirements allows better cross-version
# (Django) testing.

MySQL-python==1.2.5
pysolr>=3.3.0,<3.4.0

pytz>=2014,<2015
argparse==1.2.1
python-dateutil==2.4.1
six==1.9.0
polib==1.0.8                            # Translation file loading

Pillow>=2.7.0,<3.0.0
sorl-thumbnail>=12.4,<13.0
textract>=1.2.0,<1.3.0
Unidecode==0.04.19
parsedatetime==2.2

django-tastypie==0.13.3
django-tablib>=0.9.11,<4.0
django-crispy-forms>=1.4.0,<2.0
django-tinymce>=2.6.0,<2.7.0
django-wysiwyg>=0.7.0,<1.0
django-haystack==2.7.0
django-autoslug==1.9.3
django-modeltranslation>=0.12,<0.13
django-fsm==2.4.0


django-extensions==1.7.5
defusedxml==0.5.0
dj_database_url==0.4.2

dicttoxml==1.7.4

git+https://github.com/mPowering/orb-api-python.git@v0.1.1#orb_api

Markdown==2.6.9
enum34==1.1.6
